core: Wait for background tasks on SuperLattice destruction
Mitigates issue #6

This will wait for any background task of the stages internally used
by SuperLattice on first destruction of a super lattice which should
be fine for common usage scenarios (at worst a background task on
simulation shutdown may be waited on prematurely).

Currently the thread pool and stages are shared between any SuperLattice
instance. At some point having dedicated stages per instance would
be nice but I do not think this is a priority.